Revision 12.10.22

+++ engine/platform_win32.cpp
@@ -77,7 +77,7 @@ f64 platform_get_absolute_time() {
 }
 
 void platform_sleep(u64 ms) {
-    Sleep(ms);
+    Sleep((DWORD) ms);
 }
 
 #endif  // KPLATFORM_WINDOWS
